Copying selectCell module and changes to giles_unused
- I am going to revert this code, but don't want to lose it.
- I will be implementing an alternative solution with column selection.
This will just mark rows and columns as selected rather than individual
cells.  Tis means that you will not be able to turn on/off
arbitrary cells, instead
selecting a cell in a new column will effectively select the same column
in all currently selected rows.
This enables us to select cells in a column for all rows in the table,
which is not possible with cell selection as cell components only exist
for currently rendered rows, which, with the virtual DOM, is probably
not all rows in the table.

- Once reverted, the old code will still be available in Git history so
this copy isn't really required.   However, it's easier to make a copy for
reference whilst implementing the alternative solution.  Once this is
finished, I may delete this folder completely and just refer to git if
required.