Revert "New select_cell module" SHA 6e7f4eecdb7
- Reverting this code because we no longer require cell selection.
- I will be implementing an alternative solution with column selection.
This will just mark rows and columns as selected rather than individual
cells. This means that you will not be able to turn on/off
arbitrary cells, instead
selecting a cell in a new column will effectively select the same column
in all currently selected rows.
This enables us to select cells in a column for all rows in the table,
which is not possible with cell selection as cell components only exist
for currently rendered rows, which, with the virtual DOM, is probably
not all rows in the table.
- code for cellSelection method is temporarily being stored in giles_unused
folder for reference whilst implementing column selection.