Undoing previous change to remove tabindex from editable components.
- This will no longer be required for YoYotta purposes as we will not be
using Tabulator's editing functionality.
- I misunderstood exactily why this change caused my keyboard input
to fail.  I now understand what was happening a bit better.
- The reason that tabinput on these elements stopped keyboard input
to the table is because they allowed these elements to have keyboard
focus instead.  Once they switched to edit mode and back, the cell
had keyboard focus, not the table.  However, the keyboard event
handlers were registered to the table.  The correct fix would have
been to register keyboard handlers to the editable cells as well as the
table.   I'm not doing this now though as we no longer need this
functionality.